requirements.txt
Flask==2.2.5
prometheus_client==0.18.0
python-dotenv==1.0.1
